Module snow2d::asset

Asset management

Asset<T> is a shared reference of an asset item. Any type of Asset is stored in AssetCache.

Asset path

Asset directory is assumed to be at manifest_dir/assets. AssetKey is either of the two:

“relative/path/to/the/asset/directory”

“scheme:relative/path/to/the/scheme/directory”

TODO: <asset_dir>/schemes.txt

Serde support

Every Asset is serialized as PathBuf and deserialiezd as Asset.

WARNING: Deserialization has to be done in with_cache.

Reason: Asset is a shared pointer and we need to take care to not create duplicates. But serde doesn’t let us share states while deserialization. So we need a thread-local pointer for deserialization.

TODOs

async loading

hot reloading (tiled map, actor image, etc.)

Problems

Asset implements Deref, but they don’t implement traits that the underlying data implements

Structs

Asset
Shared ownership of an asset item. Be sure to call with_cache on deserialization.

AssetCache
All assets on memory

AssetKey
URI for an asset resolved with AssetCache::resolve. It’s a relative path from either asset directory or a directory specified by scheme.

StaticAssetKey
Prefer AssetKey::new_const

Traits

AssetItem
Data that can be used as an asset

AssetLoader
How to load an AssetItem

Functions

with_cache
Run a procedure with deserialization support for Asset

Type Definitions

Error

Result