Struct snow2d::asset::Asset

pub struct Asset<T: AssetItem> { /* fields omitted */ }

Shared ownership of an asset item

Implementations

impl<T: AssetItem> Asset<T>

pub fn is_loaded(&self) -> bool

pub fn set_preserved(&mut self, b: bool)

pub fn get<'a>(&'a self) -> Option<impl Deref<Target = T> + 'a>

Tries to get &T, fails if the asset is not loaded or failed to load

This step is for asynchrounous loading and hot reloaidng.

Unfortunatelly, the return type is not Option<&T> and doesn’t implement trait for type T. Still, you can use &*asset.get() to cast it to &T.

pub fn get_mut<'a>(&'a mut self) -> Option<impl DerefMut<Target = T> + 'a>

Tries to get &mut T, fails if the asset is not loaded or panics (Mutex under the hood)

This step is for asynchrounous loading and hot reloaidng.

Unfortunatelly, the return type is not Option<&mut T> and doesn’t implement trait for type T. Still, you can use &mut *asset.get() to cast it to &mut T.