Module snow2d::audio

Audio module

That is, soloud-rs re-exported with additional types and asset integration.

SoLoud is an easy to use, free, portable c/c++ audio engine for games.

Modules

asset
snow2d::asset integration

filter
Filter module containing all filter types and attributes

prelude
Prelude module containing all traits and error codes

src
Audio module containing all audio types

Structs

Audio
Reference-counted ownership of AudioDrop

AudioDrop
Wrapper around the Soloud native object

Handle
A wrapper around the raw handle of the audio object which can be used to adjust the parameters of the sound while it’s playing