Struct snow2d::gfx::draw::QuadParams

pub struct QuadParams {
    pub src_rect: Scaled<Rect2f>,
    pub dst_rect: Scaled<Rect2f>,
    pub origin: Vec2f,
    pub color: Color,
    pub rot: f32,
    pub flips: Flips,
    pub skew: Skew2f,
}

Full-featured geometry parameters to push a quadliteral onto Batch

Fields
src_rect: Scaled<Rect2f>dst_rect: Scaled<Rect2f>origin: Vec2f

Normalized origin

color: Colorrot: f32flips: Flipsskew: Skew2f
Implementations

impl QuadParams

pub fn reset_to_defaults(&mut self)

pub fn write_to_quad(
&self,
quad: &mut QuadData,
texture: &impl Texture2d,
flips: Flips
)

Be sure to flush Batch before running if it’s saturated.