Struct snow2d::gfx::geom2d::Rect2f

pub struct Rect2f {
    pub x: f32,
    pub y: f32,
    pub w: f32,
    pub h: f32,
}

2D rectangle

It doesn’t consider offset, rotation and scales.

Note that the x, y fields correspond top-left point. If you consider an origin at somewhere else, then those methods don’t make sense.

Fields
x: f32

Left

y: f32

Up

w: f32

Width

h: f32

Height

Implementations

impl Rect2f

pub fn new(x: f32, y: f32, w: f32, h: f32) -> Self

pub fn unit() -> Self

pub fn from_size(size: impl Into<Vec2f>) -> Self

pub fn size(&self) -> Vec2f

impl Rect2f

Primitive

pub fn left(&self) -> f32

pub fn right(&self) -> f32

pub fn up(&self) -> f32

pub fn down(&self) -> f32

pub fn set_left(&mut self, left: f32)

pub fn set_right(&mut self, right: f32)

pub fn set_up(&mut self, up: f32)

pub fn set_down(&mut self, down: f32)

pub fn left_up(&self) -> Vec2f

pub fn right_up(&self) -> Vec2f

pub fn left_down(&self) -> Vec2f

pub fn right_down(&self) -> Vec2f

pub fn set_left_up(&mut self, pos: impl Into<[f32; 2]>)

pub fn set_right_up(&mut self, pos: impl Into<[f32; 2]>)

pub fn set_left_down(&mut self, pos: impl Into<[f32; 2]>)

pub fn set_right_down(&mut self, pos: impl Into<[f32; 2]>)

impl Rect2f

More semantic

pub fn center(&self) -> Vec2f

pub fn origin_px(&self, origin: impl Into<Vec2f>) -> Vec2f

Origin in pixels from origin in normalized coordinates

pub fn offset(&self, pos: impl Into<[f32; 2]>) -> Rect2f

pub fn offset_mut(&mut self, pos: impl Into<[f32; 2]>)

pub fn set_center(&mut self, pos: impl Into<[f32; 2]>)

Sets the position of the center

pub fn set_origin(
&mut self,
pos: impl Into<[f32; 2]>,
origin: impl Into<[f32; 2]>
)

Sets the position of the origin specified with normalized coordinates

pub fn translate_mut(&mut self, offset: impl Into<Vec2f>)

Adds offset to self

pub fn clamp_x_mut(&mut self, min: f32, max: f32)

Adjusts x value assuming w < max - min (be warned that this is stupid)

pub fn clamp_y_mut(&mut self, min: f32, max: f32)

Adjusts y value assuming h < max - min (be warned that this is stupid)

impl Rect2f

Interaction

pub fn contains(&self, pos: impl Into<Vec2f>) -> bool

pub fn intersects(&self, other: &Rect2f) -> bool