Module snow2d::prelude

Primary types in snow2d

Re-exports

pub use crate::asset;

pub use crate::asset::Asset;

pub use crate::asset::AssetCache;

pub use crate::audio;

pub use crate::audio::Audio;

pub use crate::gfx::draw::*;

pub use crate::gfx::tex::*;

pub use crate::gfx::Color;

pub use crate::gfx::Snow2d;

pub use crate::gfx::WindowState;

pub use crate::input::vi;

pub use crate::ui::node::*;

pub use crate::ui::Anim;

pub use crate::ui::AnimIndex;

pub use crate::ui::Layer;

pub use crate::utils::bytemuck;

pub use crate::utils::ez;

pub use crate::utils::once_cell;

pub use crate::utils::tweak::*;

pub use crate::utils::tyobj::SerdeViaTyObj;

pub use crate::utils::tyobj::TypeObject;

pub use crate::utils::Inspect;

pub use crate::Ice;

Macros

anyhow
Construct an ad-hoc error from a string or existing non-anyhow error value.

bail
Return early with an error.

ensure
Return early with an error if a condition is not satisfied.

format_err
Construct an ad-hoc error from a string or existing non-anyhow error value.

Structs

Chain
Iterator of a chain of source errors.

Error
The Error type, a wrapper around a dynamic error type.

Input
All of the input states

Keyboard
All of the keyboard states

SoloudFlag
Soloud flags

Enums

AttenuationModel
Attenuation models

Dir4
N | E | S | W

Dir8
N | NE | E | SE | S | SW | W | NW

Key
XDL keycode

Resampler
Resampler types

Sign
Pos | Neg | Neutral

SoloudError
Soloud error types

SoloudErrorKind
Error kinds enum for SoloudError

WaveForm
Waveform types

Traits

AudioExt
Methods shared by all audio sources

Context
Provides the context method for Result.

FilterAttr
A trait applicable to all filter attributes

FilterExt
Filter creation and setting methods

FromExt
A trait defining methods of initializing audio sources from a path or memory

LoadExt
Methods for initializing sound sources from memory

Type Definitions

Result
Result<T, Error>

Derive Macros

Derivative

Inspect
Implements Inspect trait

SerdeViaTyObj
Implements SerdeViaTyObj trait

TypeObject
Implements TypeObject trait