Crate snow2d::utils::arena

Extensible generational arena for various uses. Example

Goals: Tiny code and real use. Non-goals: Super fast performance and no UB.

Features

Flexibilities:

Builtin support for distinct arena types (second type parameter of Arena<T, D, G>).

Customizable generation type (third type parameter of Arena<T, D, G>).

Unsafe goodies:

(UB) Mutable iterator of entries (Arena::bindings), rather than slot/index-based iteration.

(UB) Virtual Arena<RefCell<T>> mode: (Arena::cell).

Restrictions:

Single-threaded

Similar crates

generational_arena

thunderdome

NOTE: While arena requires extream safety, toy_arena is NOT SO TESTED (yet).

Modules

example

Example code and documentation on safety

iter

toy_arena iterator types

serde

Serde support (feature = ["serde"], which is enabled by default)

tree

Rooted tree layered on top of the generational arena

Macros

arena

Creates an arena and inserts given values. Note that you might to have to annotate your Arena<T> when the type inference doesn’t work well.

Structs

Arena

Generational arena

ArenaCell

(UB) Mutable access to multiple items in Arena at the cost of runtime check. Note that the cell does not track drops, so you can’t borrow an item once after you borrow it mutably.

Index

Slot + Gen. Takes 8 bytes for DefaultGen

Slot

Index of the backing Vec in Arena. Can be upgrade to Index by arena, but users are encouraged to perfer mutable iterators.

Traits

Gen

Generation type, one of the unsized NonZero types in std::num

Type Definitions

DefaultGen

Default generation type used by arena