Struct snow2d::utils::arena::Arena

pub struct Arena<T, D = (), G = NonZeroU32> where
    G: Gen,  { /* fields omitted */ }

Generational arena: basically a Vec, but with fixed item positions

Arena operations don’t move items. And more, each item in the arena is given “generation” value, where we can distinguish new values from old values (and see if a value is already replaced by new one).

Arena accepts three type parameters. T: item type, D: distinct type parameter for making arena newtype, and G for the internal generation type. D and G have default types, so ordinary Arena can be written as Arena<T>.

See also: crate::iter.

Implementations

impl<T, D, G> Arena<T, D, G> where
G: Gen,

pub fn new() -> Arena<T, D, G>

pub fn with_capacity(cap: usize) -> Arena<T, D, G>

pub fn len(&self) -> usize

Number of items in this arena

pub fn capacity(&self) -> usize

Capacity of the backing vec

pub fn is_empty(&self) -> bool

impl<T, D, G> Arena<T, D, G> where
G: Gen,

—– Mutations —–

pub fn insert(&mut self, data: T) -> Index<T, D, G>

pub fn clear(&mut self)

Removes all the items.

pub fn clear_and_init_generations(&mut self)

Removes all the items and resets generation of every entry.

pub fn remove(&mut self, index: Index<T, D, G>) -> Option<T>

Returns some item if the generation matchesA. Returns none on mismatch or no data

pub fn invalidate(&mut self, index: Index<T, D, G>) -> Option<Index<T, D, G>>

Returns none if the generation matches. Returns some index on mismatch or no data

pub fn invalidate_indices(
&mut self,
index: Index<T, D, G>
) -> Option<Index<T, D, G>>

Inalidates given index. Returns some index on updating the generation

pub fn replace(&mut self, index: Index<T, D, G>, new_data: T) -> Index<T, D, G>

pub fn retain<F>(&mut self, pred: F) where
F: FnMut(Index<T, D, G>, &mut T) -> bool,

Removes all items that don’t satisfy the predicate

pub unsafe fn reset_generations(&mut self)

Sets every entry’s generation to the smallest value. This operation will confuse existing indices.

impl<T, D, G> Arena<T, D, G> where
G: Gen,

—– Accessors —–

pub fn contains(&self, index: Index<T, D, G>) -> bool

pub fn get(&self, index: Index<T, D, G>) -> Option<&T>

pub fn get_mut(&mut self, index: Index<T, D, G>) -> Option<&mut T>

pub fn get2_mut(
&mut self,
ix1: Index<T, D, G>,
ix2: Index<T, D, G>
) -> Option<(&mut T, &mut T)>

Safety

Panics if the two indices point the same slot.

pub fn get3_mut(
&mut self,
ix1: Index<T, D, G>,
ix2: Index<T, D, G>,
ix3: Index<T, D, G>
) -> Option<(&mut T, &mut T, &mut T)>

Safety

Panics if any two indices point to the same slot.

pub fn upgrade(&self, slot: Slot) -> Option<Index<T, D, G>>

Upgrades slot to Index. Prefer Arena::bindings when possible

impl<T, D, G> Arena<T, D, G> where
G: Gen,

—– Iterators —–

pub fn iter(&self) -> IndexedItemIter<'_, T, D, G>

(Index, &T)

pub fn iter_mut(&mut self) -> IndexedItemIterMut<'_, T, D, G>

(Index, &mut T)

pub fn items(&self) -> ItemIter<'_, T, G>

&T

pub fn items_mut(&mut self) -> ItemIterMut<'_, T, G>

&mut T

pub fn drain(&mut self) -> Drain<'_, T, D, G>

T. Removes all items on drop

pub fn bindings(&mut self) -> EntryBindings<'_, T, D, G>

See EntryBindings and EntryBind