Trait snow2d::utils::inspect::imgui::__core::iter::Product

pub trait Product<A = Self> {
    fn product<I>(iter: I) -> Self
    where
        I: Iterator<Item = A>;
}

Trait to represent types that can be created by multiplying elements of an iterator.

This trait is used to implement Iterator::product(). Types which implement this trait can be generated by using the product() method on an iterator. Like FromIterator, this trait should rarely be called directly.

Required methods

fn product<I>(iter: I) -> Self where
I: Iterator<Item = A>,

Method which takes an iterator and generates Self from the elements by multiplying the items.