Struct snow2d::utils::inspect::imgui::ColorPicker

pub struct ColorPicker<'a, T> where
    T: 'a + AsRef<str>,  { /* fields omitted */ }

Builder for a color picker widget.

Examples

let cp = ColorPicker::new(im_str!("color_picker"), &mut color);
if cp.build(&ui) {
  println!("A color was picked");
}

Implementations

impl<'a, T> ColorPicker<'a, T> where
T: AsRef<str>,

pub fn new(label: T, value: impl Into<EditableColor<'a>>) -> ColorPicker<'a, T>

Constructs a new color picker builder.

pub fn flags(self, flags: ColorEditFlags) -> ColorPicker<'a, T>

Replaces all current settings with the given flags.

pub fn alpha(self, value: bool) -> ColorPicker<'a, T>

Enables/disables the use of the alpha component.

pub fn options(self, value: bool) -> ColorPicker<'a, T>

Enables/disables toggling of the options menu when right-clicking on inputs or the small preview.

pub fn small_preview(self, value: bool) -> ColorPicker<'a, T>

Enables/disables the colored square preview next to the inputs.

pub fn inputs(self, value: bool) -> ColorPicker<'a, T>

Enables/disables the input sliders/text widgets.

pub fn tooltip(self, value: bool) -> ColorPicker<'a, T>

Enables/disables the tooltip that appears when hovering the preview.

pub fn label(self, value: bool) -> ColorPicker<'a, T>

Enables/disables display of the inline text label (the label is in any case forwarded to the tooltip and picker).

pub fn side_preview(self, value: bool) -> ColorPicker<'a, T>

Enables/disables the bigger color preview on the right side of the picker.

pub fn alpha_bar(self, value: bool) -> ColorPicker<'a, T>

Enables/disables the vertical alpha bar/gradient in the color picker.

pub fn preview(self, preview: ColorPreview) -> ColorPicker<'a, T>

Sets the preview style.

pub fn input_mode(self, input_mode: ColorEditInputMode) -> ColorPicker<'a, T>

Sets the data format for input and output data.

pub fn display_rgb(self, value: bool) -> ColorPicker<'a, T>

Enables/disables displaying the value as RGB.

pub fn display_hsv(self, value: bool) -> ColorPicker<'a, T>

Enables/disables displaying the value as HSV.

pub fn display_hex(self, value: bool) -> ColorPicker<'a, T>

Enables/disables displaying the value as hex.

pub fn mode(self, mode: ColorPickerMode) -> ColorPicker<'a, T>

Sets the hue/saturation/value editor mode.

pub fn format(self, format: ColorFormat) -> ColorPicker<'a, T>

Sets the formatting style of color components.

pub fn reference_color(self, ref_color: &'a [f32; 4]) -> ColorPicker<'a, T>

Sets the shown reference color.

pub fn build(self, ui: &Ui<'_>) -> bool

Builds the color picker.

Returns true if the color value was changed.