Struct snow2d::utils::inspect::imgui::Context

pub struct Context { /* fields omitted */ }

An imgui-rs context.

A context needs to be created to access most library functions. Due to current Dear ImGui design choices, at most one active Context can exist at any time. This limitation will likely be removed in a future Dear ImGui version.

If you need more than one context, you can use suspended contexts. As long as only one context is active at a time, it’s possible to have multiple independent contexts.

Examples

Creating a new active context:

let ctx = imgui::Context::create();
// ctx is dropped naturally when it goes out of scope, which deactivates and destroys the
// context

Never try to create an active context when another one is active:

let ctx1 = imgui::Context::create();

let ctx2 = imgui::Context::create(); // PANIC

Suspending an active context allows you to create another active context:

let ctx1 = imgui::Context::create();
let suspended1 = ctx1.suspend();
let ctx2 = imgui::Context::create(); // this is now OK

Implementations

impl Context

pub fn create() -> Context

Creates a new active imgui-rs context.

Panics

Panics if an active context already exists

pub fn create_with_shared_font_atlas(
shared_font_atlas: Rc<RefCell<SharedFontAtlas>>
) -> Context

Creates a new active imgui-rs context with a shared font atlas.

Panics

Panics if an active context already exists

pub fn suspend(self) -> SuspendedContext

Suspends this context so another context can be the active context.

pub fn ini_filename(&self) -> Option<PathBuf>

Returns the path to the ini file, or None if not set

pub fn set_ini_filename<T>(&mut self, ini_filename: T) where
T: Into<Option<PathBuf>>,

Sets the path to the ini file (default is “imgui.ini”)

Pass None to disable automatic .Ini saving.

pub fn log_filename(&self) -> Option<PathBuf>

Returns the path to the log file, or None if not set

pub fn set_log_filename<T>(&mut self, log_filename: T) where
T: Into<Option<PathBuf>>,

Sets the log filename (default is “imgui_log.txt”).

pub fn platform_name(&self) -> Option<&str>

Returns the backend platform name, or None if not set

pub fn set_platform_name<T>(&mut self, platform_name: T) where
T: Into<Option<String>>,

Sets the backend platform name

pub fn renderer_name(&self) -> Option<&str>

Returns the backend renderer name, or None if not set

pub fn set_renderer_name<T>(&mut self, renderer_name: T) where
T: Into<Option<String>>,

Sets the backend renderer name

pub fn load_ini_settings(&mut self, data: &str)

Loads settings from a string slice containing settings in .Ini file format

pub fn save_ini_settings(&mut self, buf: &mut String)

Saves settings to a mutable string buffer in .Ini file format

pub fn set_clipboard_backend<T>(&mut self, backend: T) where
T: ClipboardBackend,

Sets the clipboard backend used for clipboard operations

impl Context

pub fn io(&self) -> &Io

Returns an immutable reference to the inputs/outputs object

pub fn io_mut(&mut self) -> &mut Io

Returns a mutable reference to the inputs/outputs object

pub fn style(&self) -> &Style

Returns an immutable reference to the user interface style

pub fn style_mut(&mut self) -> &mut Style

Returns a mutable reference to the user interface style

pub fn fonts(&mut self) -> FontAtlasRefMut<'_>

Returns a mutable reference to the font atlas.

Panics

Panics if the context uses a shared font atlas that is already borrowed

pub fn frame(&mut self) -> Ui<'_>

Starts a new frame and returns an Ui instance for constructing a user interface.

Panics

Panics if the context uses a shared font atlas that is already borrowed

impl Context

pub fn time(&self) -> f64

Returns the global imgui-rs time.

Incremented by Io::delta_time every frame.

pub fn frame_count(&self) -> i32

Returns the global imgui-rs frame count.

Incremented by 1 every frame.