Crate snow2d::utils::pool

Pool with reference-counted items

Items in the Pool will be reference-counted with strong Handles. When no Handle is referring to an item, it can be removed on synchronization, or you can handle it manually.

Note that the pool does NOT drop unreferenced items until it’s synced. Also it’s single-thread only, for no particular reason.

Modules

iter

Iterator types for the Pool

smpsc

Single-threaded mpsc backed by Rc<RefCell<VecDeque>>, just for the API

tree

Tree support

Structs

Handle

Owing index to an item in a Pool

Pool

Dynamic array with reference-counted Handles

Slot

Newtype of u32

WeakHandle

Non-owing index to an item in a Pool

Type Definitions

RefCount

Type for reference counting