Struct snow2d::utils::pool::Pool

pub struct Pool<T> { /* fields omitted */ }

Dynamic array with reference-counted Handles

Implementations

impl<T> Pool<T>

pub fn with_capacity(cap: usize) -> Self

pub fn iter(&self) -> impl Iterator<Item = &T> where
T: 'static,

Iterator of valid items in this pool

pub fn iter_mut(&mut self) -> impl Iterator<Item = &mut T> where
T: 'static,

Mutable iterator of valid items in this pool

pub fn sync_refcounts(&mut self)

Update reference counting of internal items and remove unreferenced nodes

impl<T> Pool<T>

Handle-based accessors

pub fn add(&mut self, item: impl Into<T>) -> Handle<T>

Returns a reference-counted Handle for the given item

pub fn get(&self, weak_handle: &WeakHandle<T>) -> Option<&T>

Tries to get a reference from a WeakHandle

For strong Handles, use index (pool[handle]).

pub fn get_mut(&mut self, weak_handle: &WeakHandle<T>) -> Option<&mut T>

Tries to get a mutable reference from a WeakHandle

For strong Handles, use index (pool[handle]).

impl<T> Pool<T>

Slot-based accessors

pub fn get_by_slot(&self, slot: Slot) -> Option<&T>

Retruns the item if it’s valid

pub fn get_mut_by_slot(&mut self, slot: Slot) -> Option<&mut T>

Retruns the item if it’s valid

pub fn slots(&self) -> impl Iterator<Item = Slot> + '_

pub fn enumerate_items(&self) -> impl Iterator<Item = (Slot, &T)>

Iterator of (Slot, &T)

pub fn enumerate_items_mut(&mut self) -> impl Iterator<Item = (Slot, &mut T)>

Iterator of (Slot, &mut T)