Module snow2d::utils::tyobj

Type object pattern with serde support

The idea is to have a static storage of TypeObjectMap for each TypeObject.

Structs

TypeObjectId
Id of TypeObject, which can be used to retrieve the TypeObject through global storage

TypeObjectMap
Maps TypeObjectId to TypeObject

TypeObjectStorage
Static storage of type objects

TypeObjectStorageBuilder
Utility for initializing static TypeObjectStorage

Enums

SerdeRepr
Reference | Embedded representation of a type object in field

Traits

SerdeViaTyObj
Internal utility to implement From and Into between SerdeRepr and target data type

TypeObject
Marker for data that define “type” of instances (type objects). Type objects are stored in static storage.

Derive Macros

SerdeViaTyObj
Implements SerdeViaTyObj trait

TypeObject
Implements TypeObject trait