Module snow2d::asset

Asset management

Asset types

Asset<T> is a shared reference of an asset item. Any type of Asset is stored in AssetCache.

Asset directory

Asset directory is assumed to be at manifest_dir/assets. AssetKey is a relative path from the asset directory. TODO: Add asset key with scheme: img:gorira

Serde support

Bind your AssetCache to AssetDeState while deserializing.

Every Asset is serialized as PathBuf and deserialiezd as Asset. Since Asset is a shared pointer, we need to take care to not create duplicates. But serde doesn’t let us to use states while serializing/deserializing. So we need a thread-local pointer for deserialization.

TODOs

async loading

hot reloading (tiled map, actor image, etc.)

fix static asset path

Asset implements Deref, but they don’t implement traits that the underlying data implements

Structs

Asset
Shared ownership of an asset item

AssetCache
All of the assets

AssetDeState
Deserialize assets without making duplicates using thread-local variable

AssetKey
Key to load asset

StaticAssetKey
Key to load asset (allocated statically)

Traits

AssetItem
Data that can be used as an asset

AssetLoader
How to load an AssetItem

Type Definitions

Result
std::io::Result re-exported