Trait snow2d::gfx::draw::OnSpritePush

pub trait OnSpritePush: Texture2d {
    fn init_quad(&self, builder: &mut impl QuadParamsBuilder);

    fn push_quad<Q: QuadIter>(
        &self, 
        draw: &mut DrawApiData<'_, '_, Q>, 
        flips: Flips
    )
    where
        Self: Sized,
    { ... }
}

What can be pushed onto QuadParamsBuilder by SpritePush

Required methods

fn init_quad(&self, builder: &mut impl QuadParamsBuilder)

Initializes a quad when starting to build a quad

Note that the quad is initialized to default value before this function is called.

Provided methods

fn push_quad<Q: QuadIter>(
&self,
draw: &mut DrawApiData<'_, '_, Q>,
flips: Flips
) where
Self: Sized,