Trait snow2d::prelude::AudioExt

pub unsafe trait AudioExt {

    fn default() -> Self;

    fn set_volume(&mut self, volume: f32);

    fn set_looping(&mut self, flag: bool);

    fn set_auto_stop(&mut self, flag: bool);

    fn set_3d_min_max_distance(&mut self, min_distance: f32, max_distance: f32);

    fn set_3d_attenuation(
        &mut self, 
        model: AttenuationModel, 
        rolloff_factor: f32
    );

    fn set_3d_doppler_factor(&mut self, doppler_factor: f32);

    fn set_3d_listener_relative(&mut self, flag: bool);

    fn set_3d_distance_delay(&mut self, delay: i32);

    fn set_inaudible_behavior(&mut self, must_tick: bool, kill: bool);

    fn set_loop_point(&mut self, loop_point: f64);

    fn loop_point(&self) -> f64;

    fn set_filter<F>(&mut self, filter_id: u32, filter: Option<&F>)
    where
        F: FilterExt;

    fn stop(&mut self);

    unsafe fn inner(&self) -> *mut *mut c_void;

}

Methods shared by all audio sources

Required methods

fn default() -> Self

Creates a default initialized object

fn set_volume(&mut self, volume: f32)

Sets the volume

fn set_looping(&mut self, flag: bool)

Set whether the audio is looping

fn set_auto_stop(&mut self, flag: bool)

Set auto stop

fn set_3d_min_max_distance(&mut self, min_distance: f32, max_distance: f32)

Set 3D min and max distances

fn set_3d_attenuation(&mut self, model: AttenuationModel, rolloff_factor: f32)

Set 3D attenuation

fn set_3d_doppler_factor(&mut self, doppler_factor: f32)

Set 3D doppler factor

fn set_3d_listener_relative(&mut self, flag: bool)

Set 3D listener relative

fn set_3d_distance_delay(&mut self, delay: i32)

Set 3D distance delay

fn set_inaudible_behavior(&mut self, must_tick: bool, kill: bool)

Set inaudible behavior

fn set_loop_point(&mut self, loop_point: f64)

Set a loop point

fn loop_point(&self) -> f64

Get the loop point

fn set_filter<F>(&mut self, filter_id: u32, filter: Option<&F>) where
F: FilterExt,

Set a filter, the filter_id is assigned by the developer and becomes the id for that filter, and to cancel pass None as a filter to the already assigned id

fn stop(&mut self)

Stop

unsafe fn inner(&self) -> *mut *mut c_void

Get the inner pointer

Safety

The inner pointer should be modified with care!