Struct snow2d::utils::arena::Arena

pub struct Arena<T> { /* fields omitted */ }

Container that can have elements inserted into it and removed from it.

Indices use the Index type, created by inserting values with Arena::insert.

Implementations

impl<T> Arena<T>

pub fn new() -> Arena<T>

Construct an empty arena.

pub fn with_capacity(capacity: usize) -> Arena<T>

Construct an empty arena with space to hold exactly capacity elements without reallocating.

pub fn len(&self) -> usize

Return the number of elements contained in the arena.

pub fn capacity(&self) -> usize

Return the number of elements the arena can hold without allocating, including the elements currently in the arena.

pub fn is_empty(&self) -> bool

Returns whether the arena is empty.

pub fn insert(&mut self, value: impl Into<T>) -> Index<T>

Insert a new value into the arena, returning an index that can be used to later retrieve the value.

pub fn contains(&self, index: Index<T>) -> bool

Returns true if the given index is valid for the arena.

pub fn contains_slot(&self, slot: u32) -> Option<Index<T>>

Checks to see whether a slot is occupied in the arena, and if it is, returns Some with the true Index of that slot (slot plus generation.) Otherwise, returns None.

pub fn get(&self, index: Index<T>) -> Option<&T>

Get an immutable reference to a value inside the arena by Index, returning None if the index is not contained in the arena.

pub fn get_mut(&mut self, index: Index<T>) -> Option<&mut T>

Get a mutable reference to a value inside the arena by Index, returning None if the index is not contained in the arena.

pub fn get2_mut(
&mut self,
index1: Index<T>,
index2: Index<T>
) -> (Option<&mut T>, Option<&mut T>)

Get mutable references of two values inside this arena at once by Index, returning None if the corresponding index is not contained in this arena.

Panics

This function panics when the two indices are equal (having the same slot number and generation).

pub fn remove(&mut self, index: Index<T>) -> Option<T>

Remove the value contained at the given index from the arena, returning it if it was present.

pub fn invalidate(&mut self, index: Index<T>) -> Option<Index<T>>

Invalidate the given index and return a new index to the same value. This is roughly equivalent to remove followed by insert, but much faster. If the old index is already invalid, this method returns None.

pub fn get_by_slot(&self, slot: u32) -> Option<(Index<T>, &T)>

Attempt to look up the given slot in the arena, disregarding any generational information, and retrieve an immutable reference to it. Returns None if the slot is empty.

pub fn get_by_slot_mut(&mut self, slot: u32) -> Option<(Index<T>, &mut T)>

Attempt to look up the given slot in the arena, disregarding any generational information, and retrieve a mutable reference to it. Returns None if the slot is empty.

pub fn remove_by_slot(&mut self, slot: u32) -> Option<(Index<T>, T)>

Remove an entry in the arena by its slot, disregarding any generational info. Returns None if the slot was already empty.

pub fn clear(&mut self)

Clear the arena and drop all elements.

pub fn iter(&self) -> Iter<'_, T>

Iterate over all of the indexes and values contained in the arena.

Iteration order is not defined.

pub fn iter_mut(&mut self) -> IterMut<'_, T>

Iterate over all of the indexes and values contained in the arena, with mutable access to each value.

Iteration order is not defined.

pub fn drain(&mut self) -> Drain<'_, T>

Returns an iterator that removes each element from the arena.

Iteration order is not defined.

If the iterator is dropped before it is fully consumed, any uniterated items will be dropped from the arena, and the arena will be empty. The arena’s capacity will not be changed.

pub fn drain_filter<F>(&mut self, predicate: F) -> DrainFilter<'_, T, F> where
F: FnMut(&mut T) -> bool,

Returns an iterator that removes each element from the arena if the predicate holds.

impl<T> Arena<T>

pub fn retain<F>(&mut self, f: F) where
F: FnMut(Index<T>, &mut T) -> bool,

Remove all entries in the Arena which don’t satisfy the provided predicate.

pub fn values(&self) -> ValuesIter<'_, T>

Iterate over all of the values contained in the arena.

Iteration order is not defined.

pub fn values_mut(&mut self) -> ValuesIterMut<'_, T>

Iterate over all of the values contained in the arena with mutale access.

Iteration order is not defined.