Module snow2d::utils::pool

Reference-counted pool

pool has to be used with care: Handle is a strong reference to an item. Unreference items will be deleted in next sync call.

Why not Vec<T>?

usize index for Vec<T> has some limitations:

The position of the item at the index can change on removing other items

The item at the index might be already deleted and another new item can be there (ABA problem)

The usize index can be used for unintentional Vec<T>

Pool handles these problems:

Invalidating an item don’t change other items’ positions

Handle is always valid because items have reference-counted lifetimes with them. WeakHandle identifies their interested item with generational indices.

Handles have type parameters so the chances to use incorrect Pool reduce. (This is not perfect as handles can be used for non-original pools).

Another approach would be using non-reference-counted Index in arena.

Modules

iters
Iterator types of the pool module

Structs

Handle
Owing index to an item in a Pool

Pool
Dynamic array with reference-counted Handles

Slot
Newtype of u32

WeakHandle
Non-owing index to an item in a Pool